Set up aliases for grunt tasks in package.json
Contributors shouldn't need to install grunt / grunt-cli globally
in order to run tests and/or build the project. This commit adds
the `grunt-cli` npm package to devDependencies and sets up aliases
for the main grunt tasks (default, test and build) in the "scripts"
section  of the package.json

Also update the README to use these aliases in the documentation.