Fix a regex that is almost incorrect
Regex engines evaluate 'or' clauses from left to right. If it weren't
for the start and end of line anchors, the regex would match only the
first three characters of six-letter colour codes. For instance, it
would only match `#555` in `#555bbb`. The anchors compel the engine to
keep attempting to find a match, such that it eventually matches all six
characters. The `repeat(hexCharacter, 6)` should be passed to `or`
before `repeat(hexCharacter, 3)` is, to avoid this potential issue.